requirements.txt
git+https://github.com/jay3332/discord.py
git+https://github.com/Gorialis/jishaku
aiohttp
asyncpg
fontTools
pilmoji
Pillow
python-dotenv
unidecode
humanize